+++ packages/paperclip-diff/package.json
@@ -13,6 +13,7 @@
     "execa": "^6.0.0",
     "fs-extra": "^10.0.0",
     "glob": "^7.2.0",
+    "immer": "^9.0.7",
     "ora": "5.0.0",
     "p-limit": "^2.0.0",
     "paperclip": "^18.3.2",
