+++ packages/paperclip/native/src/lib.rs
@@ -31,6 +31,7 @@ use engine::engine::Engine;
 #[global_allocator]
 static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
 
+
 extern crate web_sys;
 
 #[wasm_bindgen]
